use ostrovok api for find hotels

+++ models/point.py
@@ -2,5 +2,5 @@ from pydantic import BaseModel
 
 
 class Point(BaseModel):
-    lat: int
-    lon: int
+    lat: float
+    lon: float
